Bogus slice values in tests/mremap-expand-slice-collision.c
Commit 399cda578564 reverted slice_boundary and slice_size to values
used by 32 bit processors, on aarch64 systems. With 512MB huge page
size, this causes negative values of the address passed to mmap.

When the value of slice_boundary reaches zero in the retry loop, that
triggers an error from SELinux due to mmap_zero; commit 6b2623ce64e1
addressed a similar problem with 32 bit processors.

ASLR can cause the heap to overlap with a region mapped by mmap, which
can trigger SIGSEGV after that region is unmapped. Commit 8ee2462f3f6e
introduced a variable named heap, which might have been intended to
solve that problem, but prior to this commit that variable is unused.

This patch conditionally increases the mmap address to a value above
the heap, aligned to the huge page size, and when that's necessary it
sets the increment (slice_size) to the huge page size. For 512MB huge
pages on aarch64, the default increment is small enough to trigger
EINVAL returns from mmap.